:art: separate infrastructure package
close #3
